docker-dev/api/requirements.txt
Flask==2.0.3
Flask-Cors==3.0.10
Pillow